requirements.txt
fastapi==0.120.3
uvicorn==0.38.0
langchain==0.3.27
langchain-google-genai==2.0.8
tavily-python==0.7.12
pydantic==2.10.6
python-dotenv==1.0.0
python-multipart==0.0.6
firebase-admin==6.3.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
google-generativeai==0.8.3
requests==2.31.0
twilio==9.0.0
pyotp==2.9.0